refactor: tiers 모듈에 presentation 레이어 추가
4계층 아키텍처 구조로 리팩토링하여 presentation 레이어 도입:
- controller와 DTO를 application에서 presentation으로 이동
- presentation 레이어를 delivery mechanism 담당 계층으로 분리
- application 레이어는 순수 orchestration 계층으로 독립화
- 의존성 방향: Presentation → Application → Domain 구조 정확화
