+++ src/tiers/tiers.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 
-import { TiersController } from './application/tiers.controller';
+import { TiersController } from './presentation/tiers.controller';
 import { TiersService } from './application/tiers.service';
 
 import { TIER_REPOSITORY } from './domain/tiers.repository.interface';
